requirements.txt
playwright
pandas
openpyxl
pytesseract
Pillow